edit room context

+++ app/api/v1/hotels/[hotelId]/rooms/[roomId]/route.ts
@@ -1,4 +1,3 @@
-import type { Room } from '@prisma/client';
 import type { NextRequest } from 'next/server';
 import { NextResponse } from 'next/server';
 
@@ -7,24 +6,32 @@ import { getHotelOrThrow } from '@/lib/server/db/utils/getHotelOrThrow';
 import { getRoomOrThrow } from '@/lib/server/db/utils/getRoomOrThrow';
 import { getUserOrThrow } from '@/lib/server/db/utils/getUserOrThrow';
 import { GeneralErrors } from '@/lib/shared/constants/errors/general';
+import { RoomErrors } from '@/lib/shared/constants/errors/rooms';
 import { HttpStatus } from '@/lib/shared/constants/httpStatus';
 import { APP_ERRORS } from '@/lib/shared/errors/api/factories';
 import { withErrorHandling } from '@/lib/shared/errors/api/withErrorHandling';
 import { checkPermission } from '@/lib/shared/utils/permissions';
+import {
+  roomSelect,
+  transformRoom,
+} from '@/lib/shared/utils/transformers/transformRoom';
 import { roomUpdateSchema } from '@/lib/shared/validation/schemas';
-import type { RoomParams } from '@/types';
+import type { RoomParams, RoomWithContext } from '@/types';
 
 export const GET = withErrorHandling(
   async (req: NextRequest, { params }: { params: RoomParams }) => {
     const { hotelId: hotelIdParam, roomId } = await params;
 
     const { id: hotelId } = await getHotelOrThrow(hotelIdParam);
-    const room = await getRoomOrThrow(roomId, hotelId);
+    const room = await getRoomOrThrow(roomId);
+    if (room.hotel.id !== hotelId)
+      throw APP_ERRORS.badRequest(RoomErrors.NOT_IN_HOTEL);
+
     const { roles } = await getUserOrThrow(req);
 
     if (!checkPermission.view.room({ roles, hotelId }))
       throw APP_ERRORS.forbidden();
-    return NextResponse.json<Room>(room);
+    return NextResponse.json<RoomWithContext>(room);
   }
 );
 
@@ -32,18 +39,27 @@ export const PATCH = withErrorHandling(
   async (req: NextRequest, { params }: { params: RoomParams }) => {
     const { hotelId: hotelIdParam, roomId } = await params;
     const { id: hotelId } = await getHotelOrThrow(hotelIdParam);
-    const room = await getRoomOrThrow(roomId, hotelId);
+    const room = await getRoomOrThrow(roomId);
+
+    if (room.hotel.id !== hotelId)
+      throw APP_ERRORS.badRequest(RoomErrors.NOT_IN_HOTEL);
+
     const { roles } = await getUserOrThrow(req);
 
     if (!checkPermission.modification.room({ roles, hotelId }))
       throw APP_ERRORS.forbidden();
 
     const data = roomUpdateSchema.parse(await req.json());
+
     const updatedRoom = await prisma.room.update({
       where: { id: room.id, hotelId },
       data,
+      select: roomSelect,
     });
-    return NextResponse.json<Room>(updatedRoom);
+
+    const transformedRoom = transformRoom(updatedRoom);
+
+    return NextResponse.json<RoomWithContext>(transformedRoom);
   }
 );
 
@@ -52,8 +68,13 @@ export const DELETE = withErrorHandling(
     const { hotelId: hotelIdParam, roomId } = await params;
 
     const { id: hotelId } = await getHotelOrThrow(hotelIdParam);
-    const room = await getRoomOrThrow(roomId, hotelId);
+    const room = await getRoomOrThrow(roomId);
+
+    if (room.hotel.id !== hotelId)
+      throw APP_ERRORS.badRequest(RoomErrors.NOT_IN_HOTEL);
+
     const { roles } = await getUserOrThrow(req);
+
     if (!checkPermission.deleion.room({ roles, hotelId }))
       throw APP_ERRORS.forbidden(GeneralErrors.ACTION_DENIED);
     await prisma.room.delete({ where: { id: room.id, hotelId } });
